moved the function to a utils file within phdi.linkage
