+++ phdi/linkage/__init__.py
@@ -31,6 +31,7 @@ from phdi.linkage.link import (
 from phdi.linkage.core import BaseMPIConnectorClient
 from phdi.linkage.postgres import DIBBsConnectorClient
 from phdi.linkage.seed import convert_to_patient_fhir_resources
+from phdi.linkage.utils import datetime_to_str
 
 
 __all__ = [
@@ -66,4 +67,5 @@ __all__ = [
     "_compare_address_elements",
     "_compare_name_elements",
     "convert_to_patient_fhir_resources",
+    "datetime_to_str",
 ]
